feat: 리포트 검색 후 list 반영 컴포넌트에 연결 (#76)
